Put footer in app shell, started simplification
